refactor createJsonList

+++ src/bin.ts
@@ -2,6 +2,7 @@
 import { exec } from 'child_process';
 import express from 'express';
 import fs from 'fs';
+// import nodemon from 'nodemon';
 import path from 'path';
 import { fileURLToPath } from 'url';
 import yargs from 'yargs';
@@ -9,6 +10,7 @@ import runAutomation from './automation.js';
 
 (async () => {
   const options = yargs
+    // .usage('Usage: --includeMount <includeMount> --page <page> --port <port> --watch <watch>')
     .usage('Usage: --includeMount <includeMount> --page <page> --port <port>')
     .option('includeMount', {
       describe: 'includes the initial mount render',
@@ -23,18 +25,22 @@ import runAutomation from './automation.js';
       describe: 'port to be used for server',
       type: 'number',
     })
+    // .option('watch', {
+    //   describe: 'generate charts on every new build',
+    //   type: 'string',
+    // })
     .argv;
 
   const {
     includeMount = false,
     page,
     port = 1235,
+    // watch = '',
   } = options;
 
   const scriptPath = fileURLToPath(import.meta.url);
   const packagePath = `${scriptPath.slice(0, scriptPath.lastIndexOf('/'))}`;
   const url = `http://localhost:${port}`;
-
 
   const createJsonList = async () => {
     const jsonList: string[] = [];
@@ -42,10 +48,13 @@ import runAutomation from './automation.js';
       if (err) throw err;
       for (const file of files) {
         if (file.includes('.json')) jsonList.push(file);
+        if (file.includes('.dsv')) fs.unlink(path.join(packagePath, file), err => {
+          if (err) throw err;
+        });
       }
-      await fs.writeFile(`${packagePath}/jsonList.dsv`, jsonList.join(' '),  err => {
-        if (err) throw err;
-      });
+    });
+    await fs.writeFile(`${packagePath}/jsonList.dsv`, jsonList.join(' '),  err => {
+      if (err) throw err;
     });
   };
 
@@ -53,10 +62,9 @@ import runAutomation from './automation.js';
     await fs.readdir(packagePath, (err, files) => {
       if (err) throw err;
       for (const file of files) {
-        if (file.includes('.json') || file.includes('.dsv'))
-          fs.unlink(path.join(packagePath, file), err => {
-            if (err) throw err;
-          });
+        if (file.includes('.json')) fs.unlink(path.join(packagePath, file), err => {
+          if (err) throw err;
+        });
       }
     });
   };
@@ -81,6 +89,13 @@ import runAutomation from './automation.js';
     app.get('/', (_, res) => res.sendFile(`${packagePath}/index.html`));
     app.listen(port, () => console.log(`automation charts displaying at ${url}`));
   };
+
+  // if (watch) {
+  //   nodemon({
+  //     script: 'app.js',
+  //     ext: 'js json'
+  //   });
+  // }
 
   await deleteJsonFiles();
   await runAutomation(page, packagePath, includeMount);
